+++ u-boot-2016/lib/gl_api.c
@@ -3,6 +3,16 @@
 #include <gl_api.h>
 #include <asm/gpio.h>
 #include <fdtdec.h>
+#include <part.h>
+#include <mmc.h>
+#include <sdhci.h>
+#include <cmd_untar.h>
+
+#ifndef CONFIG_SDHCI_SUPPORT
+extern qca_mmc mmc_host;
+#else
+extern struct sdhci_host mmc_host;
+#endif
 
 DECLARE_GLOBAL_DATA_PTR;
 
@@ -163,239 +173,173 @@ void check_button_is_press(void)
 	return;
 }
 
-//该函数未调用
-int check_test()
-{
-	int ret = 0;
-	volatile unsigned char *f1f = NULL;
-	volatile unsigned char *f2i = NULL;
-	volatile unsigned char *f3r = NULL;
-	volatile unsigned char *f4s = NULL;
-	volatile unsigned char *f5t = NULL;
-	volatile unsigned char *f6t = NULL;
-	volatile unsigned char *f7e = NULL;
-	volatile unsigned char *f8s = NULL;
-	volatile unsigned char *f9t = NULL;
-
-	volatile unsigned char *s0s = NULL;
-	volatile unsigned char *s1e = NULL;
-	volatile unsigned char *s2c = NULL;
-	volatile unsigned char *s3o = NULL;
-	volatile unsigned char *s4n = NULL;
-	volatile unsigned char *s5d = NULL;
-	volatile unsigned char *s6t = NULL;
-	volatile unsigned char *s7e = NULL;
-	volatile unsigned char *s8s = NULL;
-	volatile unsigned char *s9t = NULL;
-
-	char cmd[128] = {0};
-	sprintf(cmd, "sf probe && sf read 0x44000000 0x%x 32",
-		(CONFIG_ART_START + 0x50));
-	run_command(cmd, 0);
-
-	f1f = (volatile unsigned char *)0x44000000;
-	f2i = (volatile unsigned char *)0x44000001;
-	f3r = (volatile unsigned char *)0x44000002;
-	f4s = (volatile unsigned char *)0x44000003;
-	f5t = (volatile unsigned char *)0x44000004;
-	f6t = (volatile unsigned char *)0x44000005;
-	f7e = (volatile unsigned char *)0x44000006;
-	f8s = (volatile unsigned char *)0x44000007;
-	f9t = (volatile unsigned char *)0x44000008;
-
-	s0s = (volatile unsigned char *)0x44000010;
-	s1e = (volatile unsigned char *)0x44000011;
-	s2c = (volatile unsigned char *)0x44000012;
-	s3o = (volatile unsigned char *)0x44000013;
-	s4n = (volatile unsigned char *)0x44000014;
-	s5d = (volatile unsigned char *)0x44000015;
-	s6t = (volatile unsigned char *)0x44000016;
-	s7e = (volatile unsigned char *)0x44000017;
-	s8s = (volatile unsigned char *)0x44000018;
-	s9t = (volatile unsigned char *)0x44000019;
-
-
-	if (*f1f==0x66 && *f2i==0x69 && *f3r==0x72 && *f4s==0x73 && \
-			*f5t==0x74 && *f6t==0x74 && *f7e==0x65 && *f8s==0x73 && \
-			*f9t==0x74 && \
-			*s0s==0x73 && *s1e==0x65 && *s2c==0x63 && *s3o==0x6f && \
-			*s4n==0x6e && *s5d==0x64 && *s6t==0x74 && *s7e==0x65 && \
-			*s8s==0x73 && *s9t==0x74) {
-		printf("Device haven tested, checking MAC info...\n");
-		ret = 0;
-	} else {
-		printf("Device haven't tested, please test device in calibration firmware...\n");
-		ret = -1;
-	}
-
-	return ret;
-
-}
-
-int check_config()
-{
-	int i = 0;
-	u8 addr[6];
-	u8 addr_tmp[6] = {0xff, 0xff, 0xff, 0xff, 0xff, 0xff};
-	volatile unsigned char *tmp = NULL;
-
-	char cmd[128] = {0};
-	sprintf(cmd, "sf read 0x44000000 0x%x 16",
-		CONFIG_ART_START);
-	run_command(cmd, 0);
-
-	/*check WAN mac*/
-	for (i=0; i<6; i++) {
-		tmp = (volatile unsigned char *)0x44000000 + i;
-		addr[i] = *tmp;
-	}
-	//printf("WAN: %x:%x:%x:%x:%x:%x\n", addr[0], addr[1], addr[2], addr[3], addr[4], addr[5]);
-	if (!memcmp(addr, addr_tmp, 6)) {
-		printf("Device don't have WAN MAC info, please write MAC in calibration firmware...\n");
-		return -1;
-	}
-
-	/*check LAN mac*/
-	for (i=0; i<6; i++) {
-		tmp = (volatile unsigned char *)0x44000006 + i;
-		addr[i] = *tmp;
-	}
-	//printf("LAN: %x:%x:%x:%x:%x:%x\n", addr[0], addr[1], addr[2], addr[3], addr[4], addr[5]);
-	if (!memcmp(addr, addr_tmp, 6)) {
-		printf("Device don't have LAN MAC info, please write MAC in calibration firmware...\n");
-		return -1;
-	}
-	printf("Device have MAC info, starting firmware...\n\n");
-	return 0;
-}
-
 // 只检查文件的开头几个特殊 Magic Num
 int check_fw_type(void *address) {
-	u32 *sign_flas      = (u32 *)(address + 0x5c);
-	u16 *sign_55aa      = (u16 *)(address + 0x1fe);
-	u32 *sign_doodfeed  = (u32 *)(address);
-	u32 *sign_ubi       = (u32 *)(address);
-	u32 *sign_cdt       = (u32 *)(address);
-	u32 *sign_elf       = (u32 *)(address);
-	u32 *sign_kernel6m  = (u32 *)(address + 0x600000);
-	u32 *sign_kernel12m = (u32 *)(address + 0xc00000);
+	u32 *header_magic1 = (u32 *)(address);
+	u32 *header_magic2 = (u32 *)(address + 0x4);
 
-	if (*sign_flas == 0x73616c46)
-		return FW_TYPE_QSDK;
-	else if (*sign_ubi == 0x23494255)
-		return FW_TYPE_UBI;
-	else if (*sign_doodfeed == 0xedfe0dd0 && *sign_kernel6m == 0x73717368)
-		return FW_TYPE_FACTORY_KERNEL6M;
-	else if (*sign_doodfeed == 0xedfe0dd0 && *sign_kernel12m == 0x73717368)
-		return FW_TYPE_FACTORY_KERNEL12M;
-	else if (*sign_doodfeed == 0xedfe0dd0)
-		return FW_TYPE_FIT;
-	else if (*sign_55aa == 0xaa55)
-		return FW_TYPE_EMMC;
-	else if (*sign_cdt == 0x00544443)
-		return FW_TYPE_CDT;
-	else if (*sign_elf == 0x464c457f)
-		return FW_TYPE_ELF;
-	else
-		return FW_TYPE_UNKNOWN;
+	switch (*header_magic1) {
+		case HEADER_MAGIC_CDT:
+			return FW_TYPE_CDT;
+		case HEADER_MAGIC_ELF:
+			return FW_TYPE_ELF;
+		case HEADER_MAGIC_FIT:
+			if (*((u32 *)(address + 0x5C)) == HEADER_MAGIC_JDCLOUD)
+				return FW_TYPE_JDCLOUD;
+			else if (*((u32 *)(address + 0x600000)) == HEADER_MAGIC_SQUASHFS)
+				return FW_TYPE_FACTORY_KERNEL6M;
+			else if (*((u32 *)(address + 0xC00000)) == HEADER_MAGIC_SQUASHFS)
+				return FW_TYPE_FACTORY_KERNEL12M;
+			else
+				return FW_TYPE_FIT;
+		case HEADER_MAGIC_SYSUPGRADE1:
+			if (*header_magic2 == HEADER_MAGIC_SYSUPGRADE2)
+				return FW_TYPE_SYSUPGRADE;
+			return FW_TYPE_UNKNOWN;
+		case HEADER_MAGIC_UBI:
+			return FW_TYPE_UBI;
+		default:
+			if (*((u16 *)(address + 0x1FE)) == HEADER_MAGIC_EMMC)
+				return FW_TYPE_EMMC;
+			else
+				return FW_TYPE_UNKNOWN;
+	}
 }
 
 void print_fw_type(int fw_type) {
 	printf("* The upload file type: ");
 	switch (fw_type) {
-		case FW_TYPE_NOR:
-			printf("SPI-NOR IMGAGE *");
-			break;
-		case FW_TYPE_EMMC:
-			printf("EMMC IMAGE *");
-			break;
-		case FW_TYPE_QSDK:
-			printf("QSDK FIRMWARE *");
-			break;
-		case FW_TYPE_UBI:
-			printf("UBI FIRMWARE *");
-			break;
 		case FW_TYPE_CDT:
-			printf("CDT *");
+			printf("CDT");
 			break;
 		case FW_TYPE_ELF:
-			printf("ELF *");
+			printf("ELF");
+			break;
+		case FW_TYPE_EMMC:
+			printf("EMMC IMAGE");
 			break;
 		case FW_TYPE_FACTORY_KERNEL6M:
-			printf("FACTORY FIRMWARE (KERNEL SIZE: 6MB) *");
+			printf("FACTORY FIRMWARE (KERNEL SIZE: 6MB)");
 			break;
 		case FW_TYPE_FACTORY_KERNEL12M:
-			printf("FACTORY FIRMWARE (KERNEL SIZE: 12MB) *");
+			printf("FACTORY FIRMWARE (KERNEL SIZE: 12MB)");
 			break;
 		case FW_TYPE_FIT:
-			printf("FIT IMAGE *");
+			printf("FIT IMAGE");
+			break;
+		case FW_TYPE_JDCLOUD:
+			printf("JDCLOUD OFFICIAL FIRMWARE");
+			break;
+		case FW_TYPE_SYSUPGRADE:
+			printf("SYSUPGRADE FIRMWARE");
+			break;
+		case FW_TYPE_UBI:
+			printf("UBI FIRMWARE");
 			break;
 		case FW_TYPE_UNKNOWN:
 		default:
-			printf("UNKNOWN *");
+			printf("UNKNOWN");
 	}
+	printf(" *");
 	return;
 }
-/*
 
-int auto_update_flags = 0;
+int image_greater_than_partition(char *part_name, char *file_name, ulong file_size_in_bytes) {
+    block_dev_desc_t *blk_dev;
+    disk_partition_t disk_info = {0};
+    ulong part_size_in_blocks = 0;
+    ulong file_size_in_blocks = 0;
 
-int check_network(int tryCount)
-{
-	if (tryCount <= 0) {
-		return 1;
-	}
+    blk_dev = mmc_get_dev(mmc_host.dev_num);
+    if (blk_dev == NULL)
+        return 1;
+    if (get_partition_info_efi_by_name(blk_dev, part_name, &disk_info)) {
+        printf("\n\nPartition %s not found!", part_name);
+        return 1;
+    }
+    part_size_in_blocks = (ulong)disk_info.size;
 
-	while (tryCount--) {
-		if (run_command("ping 192.168.1.2", 0) == 0) {
-			auto_update_flags = 1;
+    if (disk_info.blksz) {
+        file_size_in_blocks = file_size_in_bytes / disk_info.blksz;
+        ulong adj_size = file_size_in_bytes % disk_info.blksz;
+        if (adj_size)
+            file_size_in_blocks += 1;
+    }
+
+    if (file_size_in_blocks > part_size_in_blocks) {
+        printf("\n\n* Error: image %s size (%lu bytes) > partition %s size (%lu bytes)! *",
+			   file_name, file_size_in_bytes, part_name, part_size_in_blocks * disk_info.blksz);
+        return 1;
+    }
+
+    return 0;
+}
+
+int check_fw_compat(const int upgrade_type, const int fw_type, const ulong file_size_in_bytes) {
+	switch (upgrade_type) {
+		case WEBFAILSAFE_UPGRADE_TYPE_FIRMWARE:
+			switch (fw_type) {
+				case FW_TYPE_FACTORY_KERNEL6M:
+					return (image_greater_than_partition("0:HLOS", "kernel", 6*1024*1024) ||
+							image_greater_than_partition("rootfs", "rootfs", file_size_in_bytes - 6*1024*1024));
+				case FW_TYPE_FACTORY_KERNEL12M:
+					return (image_greater_than_partition("0:HLOS", "kernel", 12*1024*1024) ||
+							image_greater_than_partition("rootfs", "rootfs", file_size_in_bytes - 12*1024*1024));
+				case FW_TYPE_SYSUPGRADE: {
+					const void *kernel_addr, *rootfs_addr;
+					size_t kernel_size, rootfs_size;
+					if (parse_tar_image((void *)WEBFAILSAFE_UPLOAD_RAM_ADDRESS, (size_t)file_size_in_bytes,
+										&kernel_addr, &kernel_size,
+										&rootfs_addr, &rootfs_size)
+					) {
+						printf("\n\n* The upload file is NOT valid SYSUPGRADE IMAGE!! *");
+						return 1;
+					}
+					return (image_greater_than_partition("0:HLOS", "kernel", kernel_size) ||
+							image_greater_than_partition("rootfs", "rootfs", rootfs_size));
+				}
+				case FW_TYPE_JDCLOUD:
+					break;
+				default:
+					printf("\n\n* The upload file is NOT supported FIRMWARE!! *\n\n");
+					print_fw_type(fw_type);
+					return 1;
+			}
 			break;
-		}
-		udelay (1000000);
+		case WEBFAILSAFE_UPGRADE_TYPE_UBOOT:
+			if (fw_type != FW_TYPE_ELF) {
+				printf("\n\n* The upload file is NOT supported UBOOT ELF!! *\n\n");
+				print_fw_type(fw_type);
+				return 1;
+			}
+			break;
+		case WEBFAILSAFE_UPGRADE_TYPE_IMG:
+			if (fw_type != FW_TYPE_EMMC) {
+				printf("\n\n* The upload file is NOT supported EMMC IMG!! *\n\n");
+				print_fw_type(fw_type);
+				return 1;
+			}
+			break;
+		case WEBFAILSAFE_UPGRADE_TYPE_CDT:
+			if (fw_type != FW_TYPE_CDT) {
+				printf("\n\n* The upload file is NOT supported CDT!! *\n\n");
+				print_fw_type(fw_type);
+				return 1;
+			}
+			break;
+		case WEBFAILSAFE_UPGRADE_TYPE_UIMAGE:
+			if (fw_type != FW_TYPE_FIT) {
+				printf("\n\n* The upload file is NOT supported FIT uImage!! *\n\n");
+				print_fw_type(fw_type);
+				return 1;
+			}
+			break;
+		case WEBFAILSAFE_UPGRADE_TYPE_ART:
+			break;
+		default:
+			printf("\n\n* NOT supported WEBFAILSAFE UPGRADE TYPE!! *");
+			return 1;
 	}
 
-	if (auto_update_flags) {
-		return 0;
-	} else {
-		return 1;
-	}
+	return 0;
 }
-
-int auto_update_by_tftp()
-{
-	int ret = -1;
-	char buf[128] = {0};
-	const char *file_sz_str;
-	unsigned long file_size;
-
-	if (check_network(2) != 0) {
-		//printf("host no alive.\n");
-		return -1;
-	}
-
-	if (run_command("tftpboot 0x44000000 openwrt-gl-mv1000.bin", 0) == 0) {
-		file_sz_str = getenv("filesize");
-		file_size = simple_strtoul(file_sz_str, NULL, 16);
-		if(check_fw_type((void *)0x44000000)==FW_TYPE_NOR){
-		printf("\n\n****************************\n*    FIRMWARE UPGRADING    *\n* DO NOT POWER OFF DEVICE! *\n****************************\n\n");
-		sprintf(buf,
-				"sf probe && sf update 0x%lx 0x%lx 0x%lx",
-				(unsigned long int)0x44000000,
-				(unsigned long int)WEBFAILSAFE_UPLOAD_FW_ADDRESS,
-				(unsigned long int)file_size);
-		}else if(check_fw_type((void *)0x44000000)==FW_TYPE_EMMC){
-			printf("\n\n****************************\n*    FIRMWARE UPGRADING    *\n* DO NOT POWER OFF DEVICE! *\n****************************\n\n");
-			sprintf(buf,
-					"mmc dev 0 && mmc erase 0 0x109800 && mmc write 0x%lx 0x%lx 0x%lx",
-					(unsigned long int)0x44000000,
-					(unsigned long int)0x0,
-					(unsigned long int)(file_size/512+1));
-		}else{
-			return(-1);
-		}
-		return(run_command(buf, 0));
-	}
-	return ret;
-}
-
-*/