+++ u-boot-2016/net/httpd.c
@@ -82,7 +82,7 @@ void printChecksumMd5(int address, unsigned int size)
 	int i = 0;
 	u8 output[16];
 	md5_wd(buf, size, output, CHUNKSZ_MD5);
-	printf("md5 for %08x ... %08x ==> ", address, address + size);
+	printf("md5 for 0x%08x ... 0x%08x ==> ", address, address + size);
 	for (i = 0; i < 16; i++)
 		printf("%02x", output[i] & 0xFF);
 	printf("\n");
@@ -104,14 +104,14 @@ int do_http_upgrade(const ulong size, const int upgrade_type) {
 
 	// include/gl_api.h
 	// WEBFAILSAFE_UPLOAD_RAM_ADDRESS = 0x50000000 为了可以上传更大的固件，将上传地址从 0x44000000 改为 0x50000000 避免内存 crash 重启
-	// FW_TYPE_EMMC	             1 这个是 eMMC 的 GPT 分区表或镜像，只要开头有GPT信息即可
-	// FW_TYPE_QSDK	             2 这个是官方原厂固件
-	// FW_TYPE_UBI	             3 这个是 UBI 固件，eMMC 没有 UBI 固件
-	// FW_TYPE_CDT               4 这个是 CDT 文件
-	// FW_TYPE_ELF               5 这个是 ELF 文件 (除了 U-Boot 外, SBL1, QSEE, RPM, DEVCFG 也是 ELF 文件)
-	// FW_TYPE_FACTORY_KERNEL6M	 6 这个是 Factory 格式的固件 (Kernel 大小: 6MB)
-	// FW_TYPE_FACTORY_KERNEL12M 7 这个是 Factory 格式的固件 (Kernel 大小: 12MB)
-	// FW_TYPE_FIT               8 这个是 FIT Image，包括 Factory Image 和 FIT uImage
+	// FW_TYPE_EMMC	              这个是 eMMC 的 GPT 分区表或镜像，只要开头有GPT信息即可
+	// FW_TYPE_JDCLOUD	          这个是 JDCloud 官方原厂固件
+	// FW_TYPE_UBI	              这个是 UBI 固件，eMMC 没有 UBI 固件
+	// FW_TYPE_CDT                这个是 CDT 文件
+	// FW_TYPE_ELF                这个是 ELF 文件 (除了 U-Boot 外, SBL1, QSEE, RPM, DEVCFG 也是 ELF 文件)
+	// FW_TYPE_FACTORY_KERNEL6M	  这个是 Factory 格式的固件 (Kernel 大小: 6MB)
+	// FW_TYPE_FACTORY_KERNEL12M  这个是 Factory 格式的固件 (Kernel 大小: 12MB)
+	// FW_TYPE_FIT                这个是 FIT Image，包括 Factory Image 和 FIT uImage
 
 	switch (upgrade_type) {
 		case WEBFAILSAFE_UPGRADE_TYPE_FIRMWARE:
@@ -119,64 +119,40 @@ int do_http_upgrade(const ulong size, const int upgrade_type) {
 				if (fw_type == FW_TYPE_FACTORY_KERNEL6M) {
 					printf("\n\n******************************\n* FACTORY FIRMWARE UPGRADING *\n* FIRMWARE KERNEL SIZE: 6MB  *\n*  DO NOT POWER OFF DEVICE!  *\n******************************\n\n");
 					sprintf(buf,
-						"mw 0x%lx 0x00 0x200 && "
-						"mmc dev 0 && flash 0:HLOS 0x%lx 0x%lx && flash rootfs 0x%lx 0x%lx && "
-						"mmc read 0x%lx 0x622 0x200 && mw.b 0x%lx 0x00 0x1 && mw.b 0x%lx 0x00 0x1 && mw.b 0x%lx 0x00 0x1 && flash 0:BOOTCONFIG 0x%lx 0x40000 && flash 0:BOOTCONFIG1 0x%lx 0x40000",
-						// mw 0x%lx 0x00 0x200 擦除内存中上传文件后面的 512 字节，防止文件不够 512 字节写入文件后其他字符到 eMMC
-						// 其实测试不擦除文件后内存，写入一些其他字符也可以正常启动
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+size),
+						"flash 0:HLOS 0x%lx 0x%lx && flash rootfs 0x%lx 0x%lx && "
+						"bootconfig set primary",
 						// factory.bin 由 kernel + rootfs 组成，其中 kernel 固定 6MB 大小
 						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
 						(unsigned long int)0x600000,
 						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0x600000),
-						(unsigned long int)(size-0x600000),
-						// 这部分改两个 BOOTCONFIG，启动系统 0，即 rootfs
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0x80),
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0x94),
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0xA8),
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS);
+						(unsigned long int)(size-0x600000));
 				} else if (fw_type == FW_TYPE_FACTORY_KERNEL12M) {
 					printf("\n\n******************************\n* FACTORY FIRMWARE UPGRADING *\n* FIRMWARE KERNEL SIZE: 12MB *\n*  DO NOT POWER OFF DEVICE!  *\n******************************\n\n");
 					sprintf(buf,
-						"mw 0x%lx 0x00 0x200 && "
-						"mmc dev 0 && flash 0:HLOS 0x%lx 0x%lx && flash rootfs 0x%lx 0x%lx && "
-						"mmc read 0x%lx 0x622 0x200 && mw.b 0x%lx 0x00 0x1 && mw.b 0x%lx 0x00 0x1 && mw.b 0x%lx 0x00 0x1 && flash 0:BOOTCONFIG 0x%lx 0x40000 && flash 0:BOOTCONFIG1 0x%lx 0x40000",
-						// mw 0x%lx 0x00 0x200 擦除内存中上传文件后面的 512 字节，防止文件不够 512 字节写入文件后其他字符到 eMMC
-						// 其实测试不擦除文件后内存，写入一些其他字符也可以正常启动
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+size),
+						"flash 0:HLOS 0x%lx 0x%lx && flash rootfs 0x%lx 0x%lx && "
+						"bootconfig set primary",
 						// factory.bin 由 kernel + rootfs 组成，其中 kernel 固定 12MB 大小
 						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
 						(unsigned long int)0xC00000,
 						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0xC00000),
-						(unsigned long int)(size-0xC00000),
-						// 这部分改两个 BOOTCONFIG，启动系统 0，即 rootfs
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0x80),
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0x94),
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0xA8),
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS);
-				} else if (fw_type == FW_TYPE_QSDK) {
-					printf("\n\n*******************************\n* Original FIRMWARE UPGRADING *\n*   DO NOT POWER OFF DEVICE!  *\n*******************************\n\n");
+						(unsigned long int)(size-0xC00000));
+				} else if (fw_type == FW_TYPE_JDCLOUD) {
+					printf("\n\n*******************************\n* ORIGINAL FIRMWARE UPGRADING *\n*   DO NOT POWER OFF DEVICE!  *\n*******************************\n\n");
 					sprintf(buf,
-						"mmc dev 0 && imxtract 0x%lx hlos-0cc33b23252699d495d79a843032498bfa593aba && flash 0:HLOS $fileaddr $filesize && imxtract 0x%lx rootfs-f3c50b484767661151cfb641e2622703e45020fe && flash rootfs $fileaddr $filesize && imxtract 0x%lx wififw-45b62ade000c18bfeeb23ae30e5a6811eac05e2f && flash 0:WIFIFW $fileaddr $filesize && flasherase rootfs_data && "
-						"mmc read 0x%lx 0x622 0x200 && mw.b 0x%lx 0x00 0x1 && mw.b 0x%lx 0x00 0x1 && mw.b 0x%lx 0x00 0x1 && flash 0:BOOTCONFIG 0x%lx 0x40000 && flash 0:BOOTCONFIG1 0x%lx 0x40000",
-						// 官方固件本身各个固件后面有填充 0，所以不用修改上传文件后的内存
+						"imxtract 0x%lx hlos-0cc33b23252699d495d79a843032498bfa593aba && flash 0:HLOS $fileaddr $filesize && imxtract 0x%lx rootfs-f3c50b484767661151cfb641e2622703e45020fe && flash rootfs $fileaddr $filesize && imxtract 0x%lx wififw-45b62ade000c18bfeeb23ae30e5a6811eac05e2f && flash 0:WIFIFW $fileaddr $filesize && flasherase rootfs_data && "
+						"bootconfig set primary",
 						// 执行 imxtract 时不带目标地址，则不进行复制，但会修改环境变量 $fileaddr $filesize，可以直接用
 						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
 						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-						// 这部分改两个 BOOTCONFIG，启动系统 0，即 rootfs
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0x80),
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0x94),
-						(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+0xA8),
-						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
 						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS);
+				} else if (fw_type == FW_TYPE_SYSUPGRADE) {
+					printf("\n\n*********************************\n* SYSUPGRADE FIRMWARE UPGRADING *\n*   DO NOT POWER OFF DEVICE!    *\n*********************************\n\n");
+					sprintf(buf,
+						"untar 0x%lx 0x%lx && flash 0:HLOS $kernel_addr $kernel_size && flash rootfs $rootfs_addr $rootfs_size && "
+						"bootconfig set primary",
+						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
+						(unsigned long int)size);
 				} else {
-					printf("\n\n* The upload file is NOT supported FIRMWARE!! *\n\n");
 					return (-1);
 				}
 			} else {
@@ -188,15 +164,11 @@ int do_http_upgrade(const ulong size, const int upgrade_type) {
 			printf("\n\n****************************\n*     U-BOOT UPGRADING     *\n* DO NOT POWER OFF DEVICE! *\n****************************\n\n");
 			if (sfi->flash_type == SMEM_BOOT_MMC_FLASH) {
 				if (fw_type == FW_TYPE_ELF) {
-				// mw 0x%lx 0x00 0x200 擦除内存中上传文件后面的 512 字节，防止文件不够 512 字节写入文件后其他字符到 eMMC
-				// 其实测试不擦除文件后内存，写入一些其他字符也可以正常启动
-				sprintf(buf,
-					"mw 0x%lx 0x00 0x200 && mmc dev 0 && flash 0:APPSBL 0x%lx $filesize && flash 0:APPSBL_1 0x%lx $filesize",
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+size),
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS),
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS));
+					sprintf(buf,
+						"flash 0:APPSBL 0x%lx $filesize && flash 0:APPSBL_1 0x%lx $filesize",
+						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
+						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS);
 				} else {
-					printf("\n\n* The upload file is NOT supported UBOOT ELF!! *\n\n");
 					return (-1);
 				}
 			} else {
@@ -208,9 +180,8 @@ int do_http_upgrade(const ulong size, const int upgrade_type) {
 			printf("\n\n****************************\n*      ART  UPGRADING      *\n* DO NOT POWER OFF DEVICE! *\n****************************\n\n");
 			if (sfi->flash_type == SMEM_BOOT_MMC_FLASH) {
 				sprintf(buf,
-					"mw 0x%lx 0x00 0x200 && mmc dev 0 && flash 0:ART 0x%lx $filesize",
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+size),
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS));
+					"flash 0:ART 0x%lx $filesize",
+					(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS);
 			} else {
 				printf("\n\n* Update ART is NOT supported for this FLASH TYPE yet!! *\n\n");
 				return (-1);
@@ -220,13 +191,12 @@ int do_http_upgrade(const ulong size, const int upgrade_type) {
 			printf("\n\n****************************\n*      IMG  UPGRADING      *\n* DO NOT POWER OFF DEVICE! *\n****************************\n\n");
 			if (sfi->flash_type == SMEM_BOOT_MMC_FLASH) {
 				if (fw_type == FW_TYPE_EMMC) {
-				sprintf(buf,
-					"mmc dev 0 && mmc erase 0x0 0x%lx && mmc write 0x%lx 0x0 0x%lx",
-					(unsigned long int)((size-1)/512+1),
-					(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
-					(unsigned long int)((size-1)/512+1));
+					sprintf(buf,
+						"mmc erase 0x0 0x%lx && mmc write 0x%lx 0x0 0x%lx",
+						(unsigned long int)((size-1)/512+1),
+						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
+						(unsigned long int)((size-1)/512+1));
 				} else {
-					printf("\n\n* The upload file is NOT supported EMMC IMG!! *\n\n");
 					return (-1);
 				}
 			} else {
@@ -238,13 +208,11 @@ int do_http_upgrade(const ulong size, const int upgrade_type) {
 			printf("\n\n****************************\n*      CDT  UPGRADING      *\n* DO NOT POWER OFF DEVICE! *\n****************************\n\n");
 			if (sfi->flash_type == SMEM_BOOT_MMC_FLASH) {
 				if (fw_type == FW_TYPE_CDT) {
-				sprintf(buf,
-					"mw 0x%lx 0x00 0x200 && mmc dev 0 && flash 0:CDT 0x%lx $filesize && flash 0:CDT_1 0x%lx $filesize",
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS+size),
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS),
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS));
+					sprintf(buf,
+						"flash 0:CDT 0x%lx $filesize && flash 0:CDT_1 0x%lx $filesize",
+						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS,
+						(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS);
 				} else {
-					printf("\n\n* The upload file is NOT supported CDT!! *\n\n");
 					return (-1);
 				}
 			} else {
@@ -257,14 +225,12 @@ int do_http_upgrade(const ulong size, const int upgrade_type) {
 			if (fw_type == FW_TYPE_FIT) {
 				sprintf(buf,
 					"bootm 0x%lx",
-					(unsigned long int)(WEBFAILSAFE_UPLOAD_RAM_ADDRESS));
+					(unsigned long int)WEBFAILSAFE_UPLOAD_RAM_ADDRESS);
 			} else {
-				printf("\n\n* The upload file is NOT supported FIT uImage!! *\n\n");
 				return (-1);
 			}
 			break;
 		default:
-			printf("\n\n* NOT supported WEBFAILSAFE UPGRADE TYPE!! *\n\n");
 			return (-1);
 	}